requirements.txt
fastapi[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-dotenv